renamed 'string' file to 'luminaString'
